Changed filetype
The CSS file was not a .css but a .js which is super dumb, when you want CSS and not just more Javascipt. This is fix'd now. Also added a little CSS-code to check if everything works fine.
I switched to using Notepad++ instead Processing, because it seems better / more efficient to use