Added the code for a new page and fixed some bugs
Pasted some code for a Create Class-page  from another branch into the system, so it's part of the gang now.
Added some lines (with help) of PHP-code, so the different pages also llink to the CSS-document.
Also added some of the CSS-code for the new page, that wasn't already present in template.css