Fixed the getTests() function and added an updateStatus()
getTests() is now used everywhere, so no more placeholder table rows.
updateStatus() is called first thing in getTests() before we even get the info from the "tests" table in the database.
updateStatus() just checks if the current time is before,after start- an endDate, and updates the status to the appropriate value. Then the rest of getTests() runs and shows them in the correct table