Fixed code and worked on creating tests
Fixed code for creating students from a CSV-file (with name, username, password, type - being 0 for student - and which class they belong in/to)
Started working on creating tests. 10 questions per test, all with 4 answers to choose from - 1 of which being correct. Still need to get everything to the db once the test gets created.